refactor: 引用路径跟随 cloud-ai-api-format.ts 改名
